+++ components/mitsubishi_itp/mitsubishi_itp.h
@@ -14,7 +14,6 @@
 #include "mitp_mhk.h"
 #include "itp_heatpump.h"
 #include <map>
-#include <coroutine>
 
 using namespace itp_packet;
